refactor: search path resolution
- Updated TabContainer to import SearchTab from the new search components directory.
- Refactored SearchResults to use useSearchReducer instead of useSearchState.
- Created a new SearchTab component with its own hooks and logic for managing search state.
- Implemented useInitArtistPageEffect and useInitSearchStateEffect hooks for initializing search state based on URL parameters.
- Added usePreserveSearchUrlEffect to maintain the search URL state.
- Developed useSearchReducer for managing search-related state with a reducer pattern.
- Created utility functions for determining UI state and filtering artist songs by title.
- Updated context and types for search state management.
- Enhanced search hooks for fetching artist songs and search results with caching.
- Organized and modularized search-related components and hooks for better maintainability.
