+++ frontend/src/search/components/SearchTab/hooks/useInitArtistPageEffect.ts
