+++ frontend/src/search/components/SearchTab/index.ts
@@ -1,0 +1,1 @@
+export { default as SearchTab } from "./SearchTab";
