+++ frontend/src/search/components/index.ts
@@ -1,0 +1,2 @@
+export * from "./SearchTab";
+export { default as SearchResults } from "./SearchResults";
