+++ frontend/src/search/context/SearchState.types.ts
@@ -1,0 +1,7 @@
+import type { Song } from "@chordium/types";
+
+export interface SearchState {
+  artist: string;
+  song: string;
+  results: Song[];
+}
