+++ frontend/src/search/context/index.ts
@@ -1,5 +1,3 @@
-export { SearchStateContext } from "./SearchStateContext";
 export { SearchStateProvider } from "./SearchStateProvider";
 export { defaultSearchState } from "./defaultSearchState";
 export { useSearchState } from "./useSearchState";
-export * from "./SearchStateContext.types";
