+++ frontend/src/search/hooks/index.ts
@@ -1,12 +1,6 @@
 // Re-export the main hook
-export { useSearchState } from "./useSearchState";
+export * from "./useSearchReducer";
 
 // Re-export navigation hooks
 export { useArtistNavigation } from "./useArtistNavigation";
 export { useSearchRedirect } from "./use-search-redirect";
-
-// Re-export modular components for testing/debugging if needed
-export { initialSearchState } from "./useSearchState/core/initialSearchState";
-export { searchStateReducer } from "./useSearchState/core/searchStateReducer";
-export { determineUIState } from "./useSearchState/utils/determineUIState";
-export { filterArtistSongsByTitle } from "./useSearchState/utils/filterArtistSongsByTitle";
