+++ frontend/src/search/hooks/useSearchReducer/core/index.ts
@@ -1,0 +1,2 @@
+export { initialSearchState } from "./initialSearchState";
+export { searchStateReducer } from "./searchStateReducer";
