+++ frontend/src/search/hooks/useSearchReducer/handlers/index.ts
@@ -1,0 +1,2 @@
+export { useSearchFetch } from "./useSearchFetch";
+export { useArtistSongsFetch } from "./useArtistSongsFetch";
