+++ frontend/src/search/hooks/useSearchReducer/index.ts
@@ -1,0 +1,1 @@
+export { useSearchReducer } from "./useSearchReducer";
