+++ frontend/src/search/hooks/useSongActions.ts
@@ -2,7 +2,7 @@ import { useCallback } from "react";
 import { Song } from "@/types/song";
 import { useNavigate } from "react-router-dom";
 import { toSlug } from "@/utils/url-slug-utils";
-import type { UseSongActionsProps } from "@/search/types";
+import type { UseSongActionsProps } from "../types";
 
 /**
  * useSongActions provides two main actions:
